+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   ([#165]).
 - Introduce an `EventInjector` bound to a single model input ([#167]).
 - Add a possibility to inject queries via the global `Injector`.
+- Rename gRPC method `terminate` to `tear_down`
 
 # 1.0.0 (2026-02-03)